requirements.txt
click>=6.6
Fabric>=1.12.0
Flask>=0.11.1
requests>=2.11.1
simplejson>=3.10.0
gunicorn>=19.6.0
nose2>=0.6.5
Flask-WTF>=0.14.2
Flask-RESTful>=0.3.5
pymongo>=3.4.0
python-dotenv>=0.6.4
